Move application into the interface adapter layer
